requirements.txt
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==1.5.3
yfinance==0.2.18
python-binance==1.0.17
websocket-client==1.6.1
ccxt==4.0.77
schedule==1.2.0
python-dotenv==1.0.0
pytz==2023.3
matplotlib==3.7.2
plotly==5.15.0
scikit-learn==1.3.0
ta==0.10.2
python-dateutil==2.8.2 